test(android): extract session samples into debug harness
Move the Compose previews and mock session data out of the production
SessionScreen into a debug-only sample harness (SampleSessionActivity)
that renders the real screen without a live tunnel, launchable on an
emulator via the install-emulator-sample task.
