+++ kotlin/android/app/src/debug/AndroidManifest.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <!-- Debug-only launcher for SampleSessionActivity: a second app icon ("Firezone Sample") that
+         renders the session UI against mock data. Merged into the manifest for debug builds only. -->
+    <application>
+        <activity
+            android:name="dev.firezone.android.features.session.ui.compose.SampleSessionActivity"
+            android:exported="true"
+            android:label="Firezone Sample"
+            android:theme="@style/AppTheme.Base">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+    </application>
+</manifest>
